Overview: show error window if opening help fails
Displaying the help files requires a help browser. Inform the
user if that's missing. Note that we don't receive an exception
if only the help _files_ are missing. In this case the help browser
itself displays an error message in the help dialog (at least
"yelp" under GNOME behaves that way).